% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-pie-objects.R
\name{create_pie_objects}
\alias{create_pie_objects}
\title{Create a set of mesh3d objects representing a pie chart}
\usage{
create_pie_objects(labels, counts, start_degrees = 0, tilt_degrees = 0,
  height = 0.1, camera_eye = c(0, 3, 5), camera_look_at = c(0, 0, 0))
}
\arguments{
\item{labels}{the labels for the segments}

\item{counts}{the count of occurrences for each label}

\item{start_degrees}{starting angle for first pie slice (in degrees). Default: 0}

\item{tilt_degrees}{angle by which to tilt the pie towards the camera (in degrees). Default: 0}

\item{height}{height of the pie. Default: 0.1}

\item{camera_eye}{location of camera eye. Default: c(0, 3, 5)}

\item{camera_look_at}{at what point is the camera looking. Default: c(0, 0, 0)}
}
\value{
mesh3dlist of objects
}
\description{
Create a set of mesh3d objects representing a pie chart
}
