% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-threedpie.R
\name{empty}
\alias{empty}
\title{Clone of ggplot2:::empty()}
\usage{
empty(df)
}
\arguments{
\item{df}{data.frame}
}
\value{
TRUE if empty (NULL, no rows or no cols)
}
\description{
Clone of ggplot2:::empty()
}
