% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-threedpie.R
\name{geom_threedpie}
\alias{geom_threedpie}
\title{3d pie charts in ggplot2}
\usage{
geom_threedpie(mapping = NULL, data = NULL, stat = "count",
  position = "identity", ..., start_degrees = 0, tilt_degrees = 0,
  height = 0.1, darken = 0.15, camera_eye = c(0, 3, 5),
  camera_look_at = c(0, 0, 0), na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[=aes]{aes()}} or
\code{\link[=aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[=ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[=fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data.}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{...}{Other arguments passed on to [layer()]. These are
often aesthetics, used to set an aesthetic to a fixed value, like
`colour = "red"` or `size = 3`. They may also be parameters
to the paired geom/stat.}

\item{start_degrees}{starting angle for first pie slice (in degrees). Default: 0}

\item{tilt_degrees}{angle by which to tilt the pie towards the camera (in degrees). Default: 0.}

\item{height}{height of the pie. Default: 0.1}

\item{darken}{fraction by which to darken colours for the side of the pie. Default: 0.15}

\item{camera_eye}{location of camera eye. Default: c(0, 3, 5)}

\item{camera_look_at}{at what point is the camera looking. Default: c(0, 0, 0)}

\item{na.rm}{If `FALSE`, the default, missing values are removed with
a warning. If `TRUE`, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[=borders]{borders()}}.}
}
\description{
3d pie charts in ggplot2
}
\examples{
\dontrun{
  ggplot(mtcars)  +
    geom_threedpie(aes(as.factor(cyl))) +
    coord_equal() +
    theme_minimal()
}

}
