% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-threedpie.R
\docType{data}
\name{GeomThreedPie}
\alias{GeomThreedPie}
\title{Create a new ggproto object}
\arguments{
\item{_class}{Class name to assign to the object. This is stored as the class
attribute of the object. This is optional: if `NULL` (the default),
no class name will be added to the object.}

\item{_inherit}{ggproto object to inherit from. If `NULL`, don't
inherit from any object.}

\item{...}{A list of members in the ggproto object.}
}
\description{
Create a new ggproto object
}
\section{Calling methods in a parent}{

To explicitly call a methods in a parent, use
`ggproto_parent(Parent, self)`.
}

\keyword{datasets}
